fix(web): keep TMS provider server code out of client bundles
Break client import chains that pulled crowdin-provider and pg into Turbopack
by using a static CAT capability map and restoring lightweight OAuth scope modules.
